Run typescript tests without legacy-peer-deps
By downgrading jest and installing babel
