+++ package.json
@@ -14,6 +14,9 @@
     "dist",
     "src"
   ],
+  "resolutions": {
+    "@types/babel__traverse": "7.18.2"
+  },
   "scripts": {
     "test": "jest",
     "docs:generate": "typedoc",
@@ -53,6 +56,9 @@
     "qs": "^6.10.1"
   },
   "devDependencies": {
+    "@babel/core": "^7.27.1",
+    "@babel/preset-env": "^7.27.2",
+    "@types/babel__traverse": "^7.11.1",
     "@types/dotenv-flow": "^3.1.1",
     "@types/express": "^4.17.12",
     "@types/jest": "^29.5.14",
@@ -60,17 +66,18 @@
     "@types/luxon": "^1.27.0",
     "@typescript-eslint/eslint-plugin": "^4.26.0",
     "@typescript-eslint/parser": "^4.26.0",
+    "babel-jest": "^29.7.0",
     "chai": "^4.2.0",
     "eslint": "^7.27.0",
     "eslint-config-airbnb-typescript": "^12.3.1",
     "eslint-plugin-import": "^2.23.4",
-    "jest": "^29.5.0",
+    "jest": "^26.6.3",
     "jsdoc-to-markdown": "^5.0.0",
     "mocha": "^10.2.0",
     "prettier": "^3.5.3",
     "proxyquire": "^2.1.0",
     "sinon": "^7.3.2",
-    "ts-jest": "^29.3.2",
+    "ts-jest": "^26.5.6",
     "ts-node": "^10.9.2",
     "typedoc": "^0.20.36",
     "typedoc-github-wiki-theme": "^0.3.0",
