+++ tsconfig.eslint.json
@@ -1,4 +1,4 @@
 {
   "extends": "./tsconfig.json",
-  "include": ["src/", "test/"]
+  "include": ["src/", "test/", "jest.config.ts"]
 }
